Initial commit: Add Website Content Summarizer and README
This commit includes the following:

Website Content Summarizer: A Flask web application that takes a URL as input and returns a summarized version of the content found on the page using the GPT-3.5-turbo model.
README file: Provides an overview of the repository, setup instructions, and guidance on running the projects.

Signed-off-by: Hoyeon Yu, <zz31@naver.com>